Abstracts the four group access methods out of the bookmarks controller into a concern, so that we can use them in other controllers
